remove zod validation

+++ back/src/validations/github.user.schema.js
@@ -1,12 +1,13 @@
 const { z } = require("zod");
 
 const githubUserSchema = z.object({
-      id: z.number().int().nonnegative(),
-      name: z.string(),
-      mass:z.number().int().nullable(),
-      height:z.number().int().nullable(),
-      gender: z.string().nullable(),  
-   });
+  id: z.number().int().nonnegative(),
+  name: z.string(),
+  mass: z.number().nullable(),
+  height: z.number().nullable(),
+  gender: z.string().nullable(),  
+});
+
 
 const githubUserListSchema = z.array(githubUserSchema);
